+++ headless_quake2/client/cl_newfx.c
@@ -51,7 +51,7 @@ void vectoangles2 (vec3_t value1, vec3_t angles)
 	{
 	// PMM - fixed to correct for pitch of 0
 		if (value1[0])
-			yaw = (atan2(value1[1], value1[0]) * 180 / M_PI);
+			yaw = (atan2f(value1[1], value1[0]) * 180 / M_PI);
 		else if (value1[1] > 0)
 			yaw = 90;
 		else
@@ -60,8 +60,8 @@ void vectoangles2 (vec3_t value1, vec3_t angles)
 		if (yaw < 0)
 			yaw += 360;
 
-		forward = sqrt (value1[0]*value1[0] + value1[1]*value1[1]);
-		pitch = (atan2(value1[2], forward) * 180 / M_PI);
+		forward = sqrtf (value1[0]*value1[0] + value1[1]*value1[1]);
+		pitch = (atan2f(value1[2], forward) * 180 / M_PI);
 		if (pitch < 0)
 			pitch += 360;
 	}
@@ -466,9 +466,9 @@ void CL_Heatbeam (vec3_t start, vec3_t end)
 //	for (i=0 ; i<len ; i++)
 	for (i=0 ; i<len ; i+=step)
 	{
-		d = i * 0.1 - fmod(ltime,16.0)*M_PI;
-		c = cos(d)/1.75;
-		s = sin(d)/1.75;
+		d = i * 0.1 - fmodf(ltime,16.0)*M_PI;
+		c = cosf(d)/1.75;
+		s = sinf(d)/1.75;
 #ifdef DOUBLE_SCREW		
 		for (k=-1; k<2; k+=2)
 		{
@@ -558,7 +558,7 @@ void CL_Heatbeam (vec3_t start, vec3_t forward)
 	// otherwise assume SOFT
 
 	ltime = (float) cl.time/1000.0;
-	start_pt = fmod(ltime*96.0,step);
+	start_pt = fmodf(ltime*96.0,step);
 	VectorMA (move, start_pt, vec, move);
 
 	VectorScale (vec, step, vec);
@@ -583,13 +583,13 @@ void CL_Heatbeam (vec3_t start, vec3_t forward)
 			
 			p->time = cl.time;
 			VectorClear (p->accel);
-//			rot+= fmod(ltime, 12.0)*M_PI;
-//			c = cos(rot)/2.0;
-//			s = sin(rot)/2.0;
+//			rot+= fmodf(ltime, 12.0)*M_PI;
+//			c = cosf(rot)/2.0;
+//			s = sinf(rot)/2.0;
 //			variance = 0.4 + ((float)rand()/(float)RAND_MAX) *0.2;
 			variance = 0.5;
-			c = cos(rot)*variance;
-			s = sin(rot)*variance;
+			c = cosf(rot)*variance;
+			s = sinf(rot)*variance;
 			
 			// trim it so it looks like it's starting at the origin
 			if (i < 10)
@@ -661,8 +661,8 @@ void CL_Heatbeam (vec3_t start, vec3_t end)
 		VectorClear (p->accel);
 		
 		d = crand()*M_PI;
-		c = cos(d)*30;
-		s = sin(d)*30;
+		c = cosf(d)*30;
+		s = sinf(d)*30;
 
 		p->alpha = 1.0;
 		p->alphavel = -5.0 / (1+frand());
@@ -679,7 +679,7 @@ void CL_Heatbeam (vec3_t start, vec3_t end)
 /*
 
 	ltime = (float) cl.time/1000.0;
-	start_pt = fmod(ltime*16.0,step);
+	start_pt = fmodf(ltime*16.0,step);
 	VectorMA (move, start_pt, vec, move);
 
 	VectorScale (vec, step, vec);
@@ -703,11 +703,11 @@ void CL_Heatbeam (vec3_t start, vec3_t end)
 			
 			p->time = cl.time;
 			VectorClear (p->accel);
-//			rot+= fmod(ltime, 12.0)*M_PI;
-//			c = cos(rot)/2.0;
-//			s = sin(rot)/2.0;
-			c = cos(rot)/1.5;
-			s = sin(rot)/1.5;
+//			rot+= fmodf(ltime, 12.0)*M_PI;
+//			c = cosf(rot)/2.0;
+//			s = sinf(rot)/2.0;
+			c = cosf(rot)/1.5;
+			s = sinf(rot)/1.5;
 			
 			// trim it so it looks like it's starting at the origin
 			if (i < 10)
@@ -882,7 +882,7 @@ void CL_TrackerTrail (vec3_t start, vec3_t end, int particleColor)
 		p->alphavel = -2.0;
 		p->color = particleColor;
 		dist = DotProduct(move, forward);
-		VectorMA(move, 8 * cos(dist), up, p->org);
+		VectorMA(move, 8 * cosf(dist), up, p->org);
 		for (j=0 ; j<3 ; j++)
 		{
 //			p->org[j] = move[j] + crand();
