+++ headless_quake2/client/menu.c
@@ -1117,11 +1117,11 @@ static void InvertMouseFunc( void *unused )
 {
 	if ( s_options_invertmouse_box.curvalue == 0 )
 	{
-		Cvar_SetValue( "m_pitch", fabs( m_pitch->value ) );
+		Cvar_SetValue( "m_pitch", fabsf( m_pitch->value ) );
 	}
 	else
 	{
-		Cvar_SetValue( "m_pitch", -fabs( m_pitch->value ) );
+		Cvar_SetValue( "m_pitch", -fabsf( m_pitch->value ) );
 	}
 }
