+++ headless_quake2/qcommon/cmodel.c
@@ -1266,9 +1266,9 @@ void CM_RecursiveHullCheck (int num, float p1f, float p2f, vec3_t p1, vec3_t p2)
 		if (trace_ispoint)
 			offset = 0;
 		else
-			offset = fabs(trace_extents[0]*plane->normal[0]) +
-				fabs(trace_extents[1]*plane->normal[1]) +
-				fabs(trace_extents[2]*plane->normal[2]);
+			offset = fabsf(trace_extents[0]*plane->normal[0]) +
+				fabsf(trace_extents[1]*plane->normal[1]) +
+				fabsf(trace_extents[2]*plane->normal[2]);
 	}
